requirements.txt
google-genai
python-telegram-bot
python-telegram-bot